backend/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
cryptography==45.0.5
defusedxml==0.7.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.104.1
flatbuffers==25.2.10
google-auth==2.40.3
google-genai==1.25.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.2
magika==0.6.2
markdownify==1.1.0
markitdown==0.1.2
MarkupSafe==3.0.2
motor==3.3.2
mpmath==1.3.0
numpy==2.3.1
onnxruntime==1.22.1
packaging==25.0
passlib==1.7.4
pdfminer.six==20221105
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.1.0
pydantic_core==2.33.2
pymongo==4.6.0
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.27.0
sympy==1.14.0
tenacity==8.5.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
